+++ CHANGES.rst
@@ -5,7 +5,7 @@ Unreleased
 
 - Improve error message when ``FileSystemCache`` methods are called with non-str keys. :pr:`170`
 - Added ``DynamoDb`` as a cache backend :pr:`209`
-
+- Added ``MongoDb`` as a cache backend
 
 Version 0.9.0
 -------------